requirements.txt
Flask==3.1.2
Pillow==12.0.0
numpy==2.3.5
gunicorn==23.0.0
scalene==1.5.55
